Don't use relative line numbers
I was using these for a while thinking that it would be helpful to know
how many lines to tell vim to move (e.g. `5j`), but it turns out that I
can guess that easily with my eyes, and if I care to look at a line
number if is just as fast to use `gg`.