Use `dispatch` to run tests
And use `leader d` to run the current test file in `doc` formatting.
This doesn't work outside of rspec, but that's the only case I've found
it useful in anyways, so I'm ok with that.

Also, add `leader c` to copy the current spec command to the clipboard
so I can run it in a permanent tmux split, which will not send failures
to a quickfix window (sometimes this makes the failure hard to read).
This doesn't actually work how I want yet but I'm going to keep toying
around with it.