Fix python lsp and add fish prompt
I've been using nix more and it's nice to not have things like pyenv
active when I'm also in a nix shell (because it messes with the path).
It turns out I *still* needed to especially set the python path when
starting `nvim`. Also, it's nice to use `SHLVL` to tell whether I'm in
nix or not in the prompt.